# docs/swagger.yaml
definitions:
...
  dto.RefreshToken:
    properties:
      refreshToken:
        type: string
    required:
    - refreshToken
    type: object
  dto.TokenDetail:
    properties:
      accessToken:
        type: string
      accessTokenExpireTime:
        type: integer
      refreshToken:
        type: string
      refreshTokenExpireTime:
        type: integer
    type: object
...
  helper.BaseHttpResponse:
    properties:
      result: {}
      statusCode:
        type: integer
      success:
        type: boolean
    type: object
  helper.BaseHttpResponseWithError:
    properties:
      error: {}
      result: {}
      statusCode:
        type: integer
      success:
        type: boolean
    type: object
  helper.BaseHttpResponseWithValidationError:
    properties:
      result: {}
      statusCode:
        type: integer
      success:
        type: boolean
      validationErrors:
        items:
          $ref: '#/definitions/helper.ValidationError'
        type: array
    type: object
  helper.ValidationError:
    properties:
      error:
        type: string
      key:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/token/refresh/:
    post:
      consumes:
      - application/json
      description: generate and get a new access token
      parameters:
      - description: user profile update
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/dto.RefreshToken'
      produces:
      - application/json
      responses:
        "200":
          description: user profile updated
          schema:
            allOf:
            - $ref: '#/definitions/helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/dto.TokenDetail'
              type: object
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/helper.BaseHttpResponseWithValidationError'
        "406":
          description: not acceptable
          schema:
            $ref: '#/definitions/helper.BaseHttpResponseWithError'
      summary: refresh access token
      tags:
      - Token
...
securityDefinitions:
  AuthBearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"